FoodDescriptor class updated
