+++ src/main/resources/hibernate.cfg.xml
@@ -14,7 +14,7 @@
         <property name="dialect">org.hibernate.dialect.MySQLDialect</property>
 
         <!-- Fortæller Hibernate hvilken class der skal mappes -->
-        <mapping class="yolo.Descriptor"/>
+        <mapping class="yolo.FoodDescriptorModel"/>
 
 
     </session-factory>
